requirements.txt
flask
psycopg2-binary
requests
Flask-CORS